Minor change in structure for the testing
