+++ babel.config.js
@@ -7,6 +7,6 @@ module.exports = {
         chrome: '67',
         safari: '11.1',
       },
-    }]
+    }],
   ],
 };
