+++ src/components/Dashboard/MainContainer/getTaskListAsArray.js
@@ -1,0 +1,15 @@
+export default (snapValue) => {
+  const taskList = [];
+
+  Object.keys(snapValue).forEach(task => (
+    taskList.push({
+      id: task,
+      name: snapValue[task].name,
+      priority: snapValue[task].priority,
+      description: snapValue[task].description,
+      status: snapValue[task].status,
+      key: task,
+    })));
+
+  return taskList;
+};
