+++ src/components/Dashboard/TasksColumn/TasksColumn.jsx
@@ -1,6 +1,7 @@
 import React, { Component } from 'react';
 import { Container } from 'react-bootstrap';
 import './TasksColumn.scss';
+import columnTitleClass from './columnTitleClass';
 import TaskItem from '../TaskItem/TaskItem';
 
 export default class TasksColumn extends Component {
@@ -8,21 +9,6 @@ export default class TasksColumn extends Component {
     super(props);
     this.state = {};
   }
-
-  columnTitleClass = () => {
-    const { title } = this.props;
-
-    switch (title) {
-      case 'To Do':
-        return 'tasksColumnToDoTitle';
-      case 'In Progress':
-        return 'tasksColumnInProgressTitle';
-      case 'Done':
-        return 'tasksColumnDoneTitle';
-      default:
-        return 'tasksColumnInvalidTitle';
-    }
-  };
 
   render() {
     const {
@@ -44,7 +30,7 @@ export default class TasksColumn extends Component {
 
     return (
       <div className="tasksColumn rounded mb-3">
-        <h1 className={this.columnTitleClass()}>
+        <h1 className={columnTitleClass(title)}>
           {title}
         </h1>
         <Container className="taskItemsContainer">
